Fix #40: Remove symlinks when they have another target
Old behavior: When objects at the symlinks' paths exist, do nothing and
log an error. Else create the symlinks.

New behavior: When symlinks at the paths exist and point to correct
targets, do nothing. If they point to another target, delete them and
create new ones. If there are files or directories at the paths, show an
error message to the user. If there is nothing, create the symlinks.