Reactivate editor after opening output
Previous behavior: Upon initial opening of the output, the output pane
was activated, resulting in losing input when the user continued typing
after starting the compilation process. When the output was already
opened, the editor would stay activated though, thus confusing the user
with 2 different behaviors for the same action.

New behavior: The editor is always activated after opening the output.
To interact with the output view (e.g. go to page) the user has to
manually activate it (e.g. by clicking into it).