Better handle project reload
The project class is deleted if a project-related global setting
is changed in order to force its reload at the next compile/sync
command.
The reload is necessary since the class now depends on the
environment. Realoading only when a command is issued avoids
uneccessary reloads when editing the global settings.

This commit adds guards to avoid the deletion of the project class
in the middle of the execution of a command.